BACKLOG-12428 : reorganize
