+++ src/javascript/JContent/ContentTree/ContentTree.utils.js
